recovery: update packet.time_sent with pacing rate
Sent/Acked.time_sent is the timestamp when the packet
is transmitted. With the pacing rate calculation,
it might indicate the future and impact the delivery
rate esimation which uses the value. Now we update
Sent.time_sent with the timestamp after packet
scheduling based on the pacing rate.

When pacing is not used in the application, this doesn't
have an impact because the application will ignore
the pacing rate.